Update CMake configuration to compile code with C++23
This update configures cmake to instruct the compiler to use the C++23
standard. Most of the code was already compliant, but some changes to
the range_map library were required to get it to compile on C++20 and
later.

Specifically, I removed the use of `allocator::rebind` which has been
deprecated, and instead use `std::allocator_traits`. I also implemented
a small number of code changes to make the compiler a little happier,
though there is still plenty of scope to clean the code up more.